+++ lib/src/client.dart
@@ -4,6 +4,7 @@ import 'dart:io';
 import 'dart:typed_data';
 
 import 'package:dart_nats_client/dart_nats_client.dart';
+import 'package:logger/logger.dart';
 
 import 'common.dart';
 import 'message.dart';
@@ -97,8 +98,18 @@ class Client {
   /// Instance ID
   int _ssid = 0;
 
+  /// Buffer for operations
+  List<int> _operationsBuffer = [];
+
+  /// Logger instance
+  Logger _logger;
+
   /// Default constructor
-  Client() {
+  Client({Logger logger}) : _logger = logger {
+    // Check logger or init new
+    if (_logger == null) {
+      _logger = Logger(level: Level.info);
+    }
     // Set status default disconnected
     status = Status.disconnected;
     // Init healthcheck
@@ -110,7 +121,7 @@ class Client {
       {int port = 4222,
       ConnectOption connectOption,
       int timeout = 5,
-      bool retry = true,
+      int retriesCount = 5,
       int retryInterval = 10}) async {
     // Init connection controller
     _connectCompleter = Completer();
@@ -125,17 +136,23 @@ class Client {
     _port = port;
     // Check connection options validity and save
     if (connectOption != null) _connectOption = connectOption;
+    // Declare last connection exception for catch exception throw retry connect
+    Exception lastException;
 
     // Start for loop for cennection with retries
-    for (var i = 0; i == 0 || retry; i++) {
+    _logger.d("Start connecting loop with $retriesCount retries count");
+    for (var i = 0; i < retriesCount; i++) {
+      _logger.d("Connect retry #$i/$retriesCount for connect to $host:$port");
       if (i == 0) {
         // If first attempt, status - connecting
         status = Status.connecting;
+        _logger.d("Connect Status = $status");
         // Add status to health check controller
         _healthcheck.add(status);
       } else {
         // If not first attempt, set status reconnecting
         status = Status.reconnecting;
+        _logger.d("Connect Status = $status");
         // Add status to health check controller
         _healthcheck.add(status);
         // Add delay on retryInterval for next attempt
@@ -150,6 +167,7 @@ class Client {
             timeout: Duration(seconds: timeout));
         // Set status connected after socket was inited
         status = Status.connected;
+        _logger.d("Connect Status = $status");
         // Add status to health check controller
         _healthcheck.add(status);
         // Set complete status to connect controller
@@ -160,30 +178,50 @@ class Client {
         _backendSubscriptAll();
         // Clear publications buffer
         _flushPubBuffer();
-
-        _buffer = [];
+        // Clear buffer
+        _operationsBuffer = [];
+        // Start lister socket
         _socket.listen((d) {
-          _buffer.addAll(d);
-          while (_receiveState == _ReceiveState.idle && _buffer.contains(13)) {
+          _operationsBuffer.addAll(d);
+          while (_receiveState == _ReceiveState.idle &&
+              _operationsBuffer.contains(13)) {
             _processOp();
           }
         }, onDone: () {
+          // On socket close action
+          _logger.d("Socket onDone event");
           status = Status.disconnected;
+          _logger.d("Connect Status = $status");
           _healthcheck.add(status);
           _socket.close();
         }, onError: (err) {
+          // On socket error
+          _logger.e("Socket onError event: $err");
           status = Status.disconnected;
+          _logger.d("Connect Status = $status");
           _healthcheck.add(status);
           _socket.close();
         });
-        // Exit from loop on success
+        // Exit from retry loop on success
         break;
-      } catch (err) {
+      }
+      // Catch socket exceptions
+      on SocketException catch (ex) {
+        _logger.e("On connection catched socket exception: $ex");
         // Close connection
         close();
-        // Set error into Completer
-        _connectCompleter.completeError(err);
+        // Save last exception
+        lastException = ex;
       }
+    }
+    // If connection not success, save error
+    if (status != Status.connected) {
+      // Set error into Completer
+      _logger.e("After all attempts socket is not connected. "
+          "lastException: $lastException");
+      _connectCompleter.completeError(lastException);
+    } else {
+      _logger.i("Connected to '$host:$port'");
     }
 
     return _connectCompleter.future;
@@ -204,24 +242,23 @@ class Client {
     });
   }
 
-  List<int> _buffer = [];
   _ReceiveState _receiveState = _ReceiveState.idle;
   String _receiveLine1 = '';
   void _processOp() async {
     ///find endline
-    var nextLineIndex = _buffer.indexWhere((c) {
+    var nextLineIndex = _operationsBuffer.indexWhere((c) {
       if (c == 13) {
         return true;
       }
       return false;
     });
     if (nextLineIndex == -1) return;
-    var line =
-        String.fromCharCodes(_buffer.sublist(0, nextLineIndex)); // retest
-    if (_buffer.length > nextLineIndex + 2) {
-      _buffer.removeRange(0, nextLineIndex + 2);
+    var line = String.fromCharCodes(
+        _operationsBuffer.sublist(0, nextLineIndex)); // retest
+    if (_operationsBuffer.length > nextLineIndex + 2) {
+      _operationsBuffer.removeRange(0, nextLineIndex + 2);
     } else {
-      _buffer = [];
+      _operationsBuffer = [];
     }
 
     ///decode operation
@@ -276,13 +313,13 @@ class Client {
       replyTo = s[3];
       length = int.parse(s[4]);
     }
-    if (_buffer.length < length) return;
-    var payload = Uint8List.fromList(_buffer.sublist(0, length));
-    // _buffer = _buffer.sublist(length + 2);
-    if (_buffer.length > length + 2) {
-      _buffer.removeRange(0, length + 2);
+    if (_operationsBuffer.length < length) return;
+    var payload = Uint8List.fromList(_operationsBuffer.sublist(0, length));
+    // _operationsBuffer = _operationsBuffer.sublist(length + 2);
+    if (_operationsBuffer.length > length + 2) {
+      _operationsBuffer.removeRange(0, length + 2);
     } else {
-      _buffer = [];
+      _operationsBuffer = [];
     }
 
     if (_subs[sid] != null) {
@@ -312,6 +349,8 @@ class Client {
   ///publish by byte (Uint8List) return true if sucess sending or buffering
   ///return false if not connect
   bool pub(String subject, Uint8List data, {String replyTo, bool buffer}) {
+    _logger.d(
+        "Publish data '$data' to subject '$subject' with reply to '$replyTo'");
     buffer ??= defaultPubBuffer;
     if (status != Status.connected) {
       if (buffer) {
@@ -350,6 +389,7 @@ class Client {
 
   ///subscribe to subject option with queuegroup
   Subscription sub(String subject, {String queueGroup}) {
+    _logger.d("Create subsription on '$subject'");
     _ssid++;
     var s = Subscription(_ssid, subject, this, queueGroup: queueGroup);
     _subs[_ssid] = s;
@@ -370,6 +410,7 @@ class Client {
 
   ///unsubscribe
   bool unSub(Subscription s) {
+    _logger.d("Unsubscribe from '$s.subject'");
     var sid = s.sid;
 
     if (_subs[sid] == null) return false;
@@ -396,54 +437,63 @@ class Client {
     }
   }
 
-  bool _add(String str) {
-    if (_socket == null) return false; //todo throw error
+  // Send data to stream or throw exception if stream in null
+  void _add(String str) {
+    if (_socket == null)
+      throw new Exception("Can't perform _add function. Socket is closed.");
+    // Add data to socket
     _socket.add(utf8.encode(str + '\r\n'));
-    return true;
   }
 
-  bool _addByte(List<int> msg) {
-    if (_socket == null) return false; //todo throw error
-
+  // Send bytes data to stream or throw exception if stream in null
+  void _addByte(List<int> msg) {
+    if (_socket == null)
+      throw new Exception("Can't perform _add function. Socket is closed.");
+    // Add data
     _socket.add(msg);
+    // Add end of line
     _socket.add(utf8.encode('\r\n'));
-    return true;
   }
 
   final _inboxs = <String, Subscription>{};
 
   /// Request will send a request payload and deliver the response message,
   /// or an error, including a timeout if no message was received properly.
-  Future<Message> request(String subj, Uint8List data,
+  Future<Message> request(String subject, Uint8List data,
       {String queueGroup, Duration timeout}) {
+    _logger.d("Send request into subject '$subject' "
+        "to queue group '$queueGroup' "
+        " with data '$data'"
+        " and timeout $timeout");
     timeout ??= Duration(seconds: 2);
     data ??= Uint8List(0);
 
-    if (_inboxs[subj] == null) {
+    if (_inboxs[subject] == null) {
       var inbox = newInbox();
-      _inboxs[subj] = sub(inbox, queueGroup: queueGroup);
+      _inboxs[subject] = sub(inbox, queueGroup: queueGroup);
     }
 
     // TODO timeout
     // TODO refactor
-    var stream = _inboxs[subj].getStream();
+    var stream = _inboxs[subject].getStream();
     var respond = stream.take(1).single;
     // Publish reply
-    pub(subj, data, replyTo: _inboxs[subj].subject);
+    pub(subject, data, replyTo: _inboxs[subject].subject);
 
     return respond;
   }
 
   /// requestString() helper to request()
-  Future<Message> requestString(String subj, String data,
+  Future<Message> requestString(String subject, String data,
       {String queueGroup, Duration timeout}) {
     data ??= '';
-    return request(subj, Uint8List.fromList(data.codeUnits),
+    return request(subject, Uint8List.fromList(data.codeUnits),
         queueGroup: queueGroup, timeout: timeout);
   }
 
   /// Close connection to NATS server unsub to server
   void close() {
+    _logger.i("Close connection to '$_host:$_port'");
     // Clear backend subscriptions
     _backendSubscriptAll();